some format fixes according to pull request comments

+++ src/airmap/rest/glib/communicator.cpp
@@ -120,7 +120,7 @@ void airmap::rest::glib::Communicator::send_udp(const std::string& host, std::ui
                 if (socket)
                   g_object_unref(socket);
               }};
-          if (static_cast<int>(body.size()) ==
+          if (static_cast<ssize_t>(body.size()) ==
               g_socket_send_to(s.get(), sa.get(), body.c_str(), body.size(), nullptr, &error)) {
             log_.infof(component, "successfully sent UDP packet");
           } else if (error) {
